requirements.txt
xlrd==2.0.1
pandas<1.5.1
Pillow==9.3.0
plotly==5.11.0
streamlit==1.14.1
